user_management/requirements.txt
asgiref==3.8.1
Django==4.2.16
djangorestframework==3.15.2
djongo==1.3.7
mongoengine==0.29.1
pymongo==3.11.4
pytz==2024.2
sqlparse==0.5.2
typing_extensions==4.12.2
python-dotenv==1.0.1
